+++ src/common/structs/mod.rs
@@ -8,8 +8,9 @@ pub mod magazines;
 pub mod manga;
 pub mod people;
 pub mod producers;
+pub mod recommendation;
 pub mod reviews;
 pub mod schedule;
 pub mod season;
 pub mod users;
-pub mod recommendation;
+pub mod watch;
